% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-loading.R
\name{convert_epa_to_multinomial}
\alias{convert_epa_to_multinomial}
\title{Convert EPA Lognormal to Multinomial}
\usage{
convert_epa_to_multinomial(datmatx_epa, epa_as_col, epa_lognormal_sdlog)
}
\arguments{
\item{datmatx_epa}{EPA data frame}

\item{epa_as_col}{Column name for EPA arsenic meanlog}

\item{epa_lognormal_sdlog}{Standard deviation for lognormal distribution}
}
\value{
Matrix of multinomial probabilities
}
\description{
Convert EPA Lognormal to Multinomial
}
\keyword{internal}
