% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_geographic_ids}
\alias{format_geographic_ids}
\title{Format Geographic Identifiers}
\usage{
format_geographic_ids(data_frame, geod_col)
}
\arguments{
\item{data_frame}{Data frame to format}

\item{geod_col}{Column name for geographic identifier}
}
\value{
Data frame with formatted geographic IDs
}
\description{
Format Geographic Identifiers
}
\keyword{internal}
