% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-loading.R
\name{generate_default_epa_cutoffs}
\alias{generate_default_epa_cutoffs}
\title{Generate Default EPA Cutoffs Based on USGS Categories}
\usage{
generate_default_epa_cutoffs(n_categories)
}
\arguments{
\item{n_categories}{Number of USGS categories}
}
\value{
Numeric vector of EPA cutoffs
}
\description{
Generate Default EPA Cutoffs Based on USGS Categories
}
\keyword{internal}
