% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-loading.R
\name{generate_logarithmic_cutoffs}
\alias{generate_logarithmic_cutoffs}
\title{Generate Logarithmic EPA Cutoffs for Unusual Category Counts}
\usage{
generate_logarithmic_cutoffs(n_categories)
}
\arguments{
\item{n_categories}{Number of categories}
}
\value{
Numeric vector of logarithmically spaced cutoffs
}
\description{
Generate Logarithmic EPA Cutoffs for Unusual Category Counts
}
\keyword{internal}
