% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impute_additional_variables}
\alias{impute_additional_variables}
\title{Impute Additional Variables Using MICE}
\usage{
impute_additional_variables(
  imputed_datasets,
  impute_vars,
  births,
  mice_covs,
  mice_m,
  mice_maxit,
  mice_method,
  seed,
  check_birth_cols = TRUE
)
}
\arguments{
\item{imputed_datasets}{List of datasets with arsenic exposure already
imputed}

\item{impute_vars}{Character vector of variables to impute}

\item{births}{Original birth data frame}

\item{mice_covs}{Character vector of covariates for MICE}

\item{mice_m}{Number of MICE imputations}

\item{mice_maxit}{Maximum MICE iterations}

\item{mice_method}{MICE imputation method}

\item{seed}{Random seed}

\item{check_birth_cols}{Logical indicating whether to check if
impute_vars exist in the birth data}
}
\value{
List of datasets with additional variables imputed
}
\description{
Internal function to impute missing values in additional variables using
MICE.
}
\keyword{internal}
