% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-loading.R
\name{load_usgs_data}
\alias{load_usgs_data}
\title{Load USGS Data}
\usage{
load_usgs_data(as_usgs_prob_csv, as_conc_cols, pop_well_col)
}
\arguments{
\item{as_usgs_prob_csv}{Path to USGS CSV file}

\item{as_conc_cols}{Column names for concentration probabilities}

\item{pop_well_col}{Column name for population weights}
}
\value{
List with USGS data, probability matrix, and cell populations
}
\description{
Load USGS Data
}
\keyword{internal}
