% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensAnalysis.R
\name{perform_sensitivity_analysis}
\alias{perform_sensitivity_analysis}
\title{Perform Sensitivity Analysis}
\usage{
perform_sensitivity_analysis(
  ndraws = 10,
  as_prob_csv,
  birth_data_txt,
  output_dir,
  targets = c("OEGEST", "BWT"),
  as_cat_label = c("As<5", "As5-10", "As10+"),
  drop_as_cat_label_reg = c("As<5"),
  seed = 12345
)
}
\arguments{
\item{ndraws}{Number of draws for the simulation.}

\item{as_prob_csv}{A CSV file containing the estimated probability of
arsenic levels.}

\item{birth_data_txt}{A TXT file containing the birth data.}

\item{output_dir}{The output directory.}

\item{targets}{A vector of target variables.}

\item{as_cat_label}{A vector of arsenic categories generated by the random
forest model.}

\item{drop_as_cat_label_reg}{A vector of arsenic categories to drop from the
regression analysis.}

\item{seed}{The seed for the random number generator.}
}
\value{
A list of data frames with the imputed arsenic levels.
}
\description{
This function performs sensitivity analysis on NAPHSIS data.
}
