% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwArsenic.R
\name{regression_analysis}
\alias{regression_analysis}
\title{Perform Regression Analysis}
\usage{
regression_analysis(
  births,
  targets,
  as_cat_label,
  as_level_col,
  regression_formula,
  output_dir
)
}
\arguments{
\item{births}{A data frame containing the birth data.}

\item{targets}{A vector of target variables.}

\item{as_cat_label}{A vector of arsenic categories generated by the random
forest model.}

\item{as_level_col}{The column name for the arsenic level.}

\item{regression_formula}{The regression formula.}

\item{output_dir}{The output directory.}
}
\value{
A list of data frames with the regression results.
}
\description{
This function performs regression analysis on NAPHSIS data.
}
