% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_arsenic_categories}
\alias{validate_arsenic_categories}
\title{Validate Arsenic Category Parameters}
\usage{
validate_arsenic_categories(as_cat_label, drop_as_cat_label_reg)
}
\description{
Validate Arsenic Category Parameters
}
\keyword{internal}
