% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-loading.R
\name{validate_category_alignment}
\alias{validate_category_alignment}
\title{Validate Category Alignment Between USGS and EPA}
\usage{
validate_category_alignment(n_usgs_categories, epa_cutoffs)
}
\arguments{
\item{n_usgs_categories}{Number of USGS categories}

\item{epa_cutoffs}{EPA cutoff values}
}
\value{
TRUE if aligned, otherwise stops with error
}
\description{
Validate Category Alignment Between USGS and EPA
}
\keyword{internal}
