% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_column_names}
\alias{validate_column_names}
\title{Validate Column Name Parameters}
\usage{
validate_column_names(
  as_level_col,
  rucc_col,
  geoid_col,
  pop_well_col,
  birth_county_col,
  epa_as_mean_col,
  epa_as_sd_col,
  epa_pwell_col
)
}
\description{
Validate Column Name Parameters
}
\keyword{internal}
