% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-loading.R
\name{validate_data_alignment}
\alias{validate_data_alignment}
\title{Validate Data Alignment}
\usage{
validate_data_alignment(datmatx_usgs, datmatx_epa)
}
\arguments{
\item{datmatx_usgs}{USGS data frame}

\item{datmatx_epa}{EPA data frame}
}
\description{
Validate Data Alignment
}
\keyword{internal}
