% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_epa_cutoff_params}
\alias{validate_epa_cutoff_params}
\title{Validate EPA Cutoffs Parameter}
\usage{
validate_epa_cutoff_params(epa_cutoffs, as_conc_cols)
}
\description{
Validate EPA Cutoffs Parameter
}
\keyword{internal}
