% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_epa_sd_values}
\alias{validate_epa_sd_values}
\title{Validate EPA Standard Deviation Values}
\usage{
validate_epa_sd_values(as_epa_prob_csv, epa_as_sd_col)
}
\description{
Validate EPA Standard Deviation Values
}
\keyword{internal}
