% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_file_contents}
\alias{validate_file_contents}
\title{Validate File Contents and Column Existence}
\usage{
validate_file_contents(
  birth_data_txt,
  as_usgs_prob_csv,
  as_epa_prob_csv,
  rucc_col,
  birth_county_col,
  geoid_col,
  as_conc_cols,
  pop_well_col,
  epa_as_mean_col,
  epa_as_sd_col,
  epa_pwell_col
)
}
\description{
Validate File Contents and Column Existence
}
\keyword{internal}
