% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_file_paths}
\alias{validate_file_paths}
\title{Validate File Paths}
\usage{
validate_file_paths(birth_data_txt, as_usgs_prob_csv, as_epa_prob_csv)
}
\description{
Validate File Paths
}
\keyword{internal}
