% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_imputation_fallback}
\alias{validate_imputation_fallback}
\title{Validate Imputation Fallback Parameter}
\usage{
validate_imputation_fallback(apply_imputation_fallback)
}
\description{
Validate Imputation Fallback Parameter
}
\keyword{internal}
