% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{validate_imputed_datasets}
\alias{validate_imputed_datasets}
\title{Validate Imputed Datasets}
\usage{
validate_imputed_datasets(imputed_datasets, as_level_col, as_cat_label)
}
\arguments{
\item{imputed_datasets}{List of imputed datasets}

\item{as_level_col}{Column name for arsenic level variable}

\item{as_cat_label}{Character vector of valid arsenic category labels}
}
\description{
Internal function to validate that arsenic exposure imputation was
successful.
}
\keyword{internal}
