% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_numeric_params}
\alias{validate_numeric_params}
\title{Validate Numeric Parameters}
\usage{
validate_numeric_params(ndraws, mice_m, mice_maxit, seed)
}
\description{
Validate Numeric Parameters
}
\keyword{internal}
