% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_regression_formula}
\alias{validate_regression_formula}
\title{Validate Regression Formula}
\usage{
validate_regression_formula(regression_formula, as_level_col)
}
\description{
Validate Regression Formula
}
\keyword{internal}
