% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_usgs_columns_param}
\alias{validate_usgs_columns_param}
\title{Validate USGS Columns Parameter}
\usage{
validate_usgs_columns_param(usgs_columns)
}
\description{
Validate USGS Columns Parameter
}
\keyword{internal}
