% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_all_inputs}
\alias{validate_all_inputs}
\title{Validate Input Parameters for Sensitivity Analysis}
\usage{
validate_all_inputs(
  birth_data_txt = NULL,
  as_usgs_prob_csv = NULL,
  as_epa_prob_csv = NULL,
  ndraws = NULL,
  regression_formula = NULL,
  targets = NULL,
  impute_vars = NULL,
  output_dir = NULL,
  mice_m = NULL,
  mice_maxit = NULL,
  mice_method = NULL,
  epa_as_mean_col = NULL,
  epa_as_sd_col = NULL,
  seed = NULL,
  as_cat_label = NULL,
  as_level_col = "AsLevel",
  drop_as_cat_label_reg = NULL,
  epa_cutoffs = NULL,
  usgs_columns = NULL,
  validate_cutoffs = TRUE,
  columns_to_select = NULL,
  rucc_col = NULL,
  geoid_col = NULL,
  as_conc_cols = NULL,
  pop_well_col = NULL,
  birth_county_col = NULL,
  epa_pwell_col = NULL,
  mice_covs = NULL,
  apply_imputation_fallback = TRUE
)
}
\arguments{
\item{...}{All parameters to validate}
}
\value{
NULL (throws error if validation fails)
}
\description{
Validate Input Parameters for Sensitivity Analysis
}
\keyword{internal}
